Add licenses and notice for 3rd-party dependencies

+++ package.sh
@@ -15,7 +15,7 @@ if test "$(uname)" = "Darwin"; then
     PACKAGE_NAME=aliyun-mq-darwin-cpp-sdk
 fi
 
-# this script must execute on graalvm-rocketmq-client4cpp dirname
+# This script must execute on project home dir
 full_path=`pwd`
 base_dir_name=`basename $full_path`
 if [[ ${base_dir_name} != ${HOME_PATH} ]];then
@@ -23,15 +23,20 @@ if [[ ${base_dir_name} != ${HOME_PATH} ]];then
   exit -1
 fi
 
+# Step 1, check license output
 
-# setp1 check headfile output
+if [[ ! -e "licenses" ]]; then
+  echo "Licenses path licenses does not exist, please check."
+  exit -1
+fi
+# Step 2, check headfile output
 
 if [[ ! -e "src/main/cpp/include" ]]; then
   echo "Head file path src/main/cpp/include does not exist, please check."
   exit -2
 fi
 
-# setp2 check lib/libonsclient4cpp.so and build/librocketmq_client_core.so
+# Step 3, check lib/libonsclient4cpp.so and build/librocketmq_client_core.so
 if [[ ! -e "build/${CORE_LIBRARY_NAME}.${LIBRARY_SUFFIX}" ]];then
   echo "Core library build/${CORE_LIBRARY_NAME}.${LIBRARY_SUFFIX} not exits, please execute mvn install"
   exit -3
@@ -42,13 +47,19 @@ echo "ONS API Library ${API_LIBRARY_NAME}.${LIBRARY_SUFFIX} not exits, please go
 exit -3
 fi
 
-
+# Step 4, check demos
 if [[ ! -e "src/main/cpp/demos" ]];then
   echo "Demo path src/main/cpp/demos does not exists, please check."
   exit -4
 fi
 
-# setp5 aliyun-mq-linux-cpp-sdk
+# Step 5, check docs
+if [[ ! -e "doc" ]];then
+  echo "doc path doc does not exists, please check."
+  exit -5
+fi
+
+# Step 6 remove old package path
 if [[ -e ${PACKAGE_NAME} ]] || [[ -e ${PACKAGE_NAME}.tar.gz ]];then
   echo "find old package dir, we will delete this package file."
   rm -rf ${PACKAGE_NAME} &>/dev/null
@@ -72,6 +83,9 @@ echo "Copy demo make file................."
 cp -f src/main/cpp/demos/CMakeLists.Release ${PACKAGE_NAME}/demos/CMakeLists.txt
 echo "Copy changlog................."
 cp -f doc/changelog ${PACKAGE_NAME}/doc/
+echo "Copy license and notice file................."
+cp -f licenses/LICENSE-BIN ${PACKAGE_NAME}/LICENSE
+cp -f licenses/NOTICE-BIN ${PACKAGE_NAME}/NOTICE
 echo "========================packaging end ================================================"
